+++ src/utils/content-extractor.ts
@@ -8,10 +8,10 @@ import { debugLog } from './debug';
 import dayjs from 'dayjs';
 
 import { AnyHighlightData } from './highlighter';
-import { processForLoop } from './tags/for_loop';
-import { processSelector } from './tags/selector';
-import { processSchema } from './tags/schema';
-import { processPrompt } from './tags/prompt';
+import { processForLoop } from './tags/for';
+import { processSelector } from './variables/selector';
+import { processSchema } from './variables/schema';
+import { processPrompt } from './variables/prompt';
 
 export function extractReadabilityContent(doc: Document): ReturnType<Readability['parse']> | null {
 	try {
@@ -27,7 +27,7 @@ export function extractReadabilityContent(doc: Document): ReturnType<Readability
 type LogicHandler = {
 	type: string;
 	regex: RegExp;
-	process: (match: RegExpExecArray, variables: { [key: string]: any }, currentUrl: string) => Promise<string>;
+	process: (match: RegExpExecArray, variables: { [key: string]: any }, currentUrl: string, processLogicStructures: (text: string, variables: { [key: string]: any }, currentUrl: string) => Promise<string>) => Promise<string>;
 };
 
 // Define logic handlers
@@ -35,22 +35,21 @@ const logicHandlers: LogicHandler[] = [
 	{
 		type: 'for',
 		regex: /{%\s*for\s+(\w+)\s+in\s+([\w:@]+)\s*%}([\s\S]*?){%\s*endfor\s*%}/g,
-		process: async (match, variables, currentUrl) => {
-			const [fullMatch, iteratorName, arrayName, loopContent] = match;
-			return processForLoop(fullMatch, variables, currentUrl);
+		process: async (match, variables, currentUrl, processLogicStructures) => {
+			return processForLoop(match, variables, currentUrl, processLogicStructures);
 		}
 	},
 	// Add more logic handlers
 ];
 
 // Function to process logic structures
-async function processLogicStructures(text: string, variables: { [key: string]: any }, currentUrl: string): Promise<string> {
+export async function processLogicStructures(text: string, variables: { [key: string]: any }, currentUrl: string): Promise<string> {
 	let processedText = text;
 	
 	for (const handler of logicHandlers) {
 		let match;
 		while ((match = handler.regex.exec(processedText)) !== null) {
-			const result = await handler.process(match, variables, currentUrl);
+			const result = await handler.process(match, variables, currentUrl, processLogicStructures);
 			processedText = processedText.substring(0, match.index) + result + processedText.substring(match.index + match[0].length);
 			handler.regex.lastIndex = match.index + result.length;
 		}
@@ -61,11 +60,8 @@ async function processLogicStructures(text: string, variables: { [key: string]: 
 
 // Main function to compile the template
 export async function compileTemplate(tabId: number, text: string, variables: { [key: string]: any }, currentUrl: string): Promise<string> {
-	// Process logic structures
-	const processedText = await processLogicStructures(text, variables, currentUrl);
-	
 	// Process other variables and filters
-	return await processVariables(tabId, processedText, variables, currentUrl);
+	return await processVariables(tabId, text, variables, currentUrl);
 }
 
 // Function to process variables and apply filters
