+++ src/utils/variables/schema.ts
